feat: generalize service so it can compare any two snapshots
